pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vlm-kie"
version = "0.1.0"
description = "VLM Key Information Extraction benchmark for scanned invoices"
requires-python = ">=3.11,<3.13"

dependencies = [
    "ollama>=0.3.0",
    "torch>=2.3.0,<2.5",
    "torchvision>=0.18.0",
    "transformers>=4.52.0",
    "accelerate>=0.34.0",
    "Pillow>=10.0.0",
    "datasets>=2.20.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "rapidfuzz>=3.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
paddle = [
    "paddlepaddle>=3.0.0",
    "paddleocr>=2.9.0",
    "paddlex>=3.4.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/vlm_kie"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
